pyproject.toml
[tool.poetry]
name = "omnyorder"
version = "1.0.0"
description = "Audio transcription and text-to-speech web application with FastAPI backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "app.py"}]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
deepgram-sdk = "^4.8.1"

# Optional dependencies for AI services
# openai = "^1.3.0"  # For Whisper API
# openai-whisper = "^20231117"  # For local Whisper
# elevenlabs = "^0.2.27"  # For Eleven Labs TTS
elevenlabs = "^2.20.1"
openai = "^1.92.0"
chromadb = "^1.2.1"
vapi-server-sdk = "^1.7.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
ignore = ["E501"]

